Self-Awareness-Workflow/requirements.txt
joblib==1.4.2
numpy==1.26.3
pandas==2.2.2
plotly==5.19.0
PyQt6==6.7.1
pyreadr==0.5.2
scikit-learn==1.5.1
torch==2.1.2
tqdm==4.66.5
pickle==4.0
kafka-python